app/pyproject.toml
[tool.poetry]
name = "mass-price-regression"
version = "0.1.0"
description = ""
authors = ["akennouche <akennouche@meilleurtaux.com>"]
readme = "README.md"

packages = [
    { include = "mass_price_regression", from="./src/" },
    { include = "mass_price_regression/**/*.py", from="./src/" },
]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "1.3.5"
loguru = "0.6.0"
scikit-learn = "*"
xgboost = "1.6.1"
streamlit = "1.28.2"
flask = "^3.0.0"
pytest = "^7.4.3"
gunicorn = "^21.2.0"
google-cloud-bigquery = "^3.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
